test: split and expand integration tests for address_manager and registration (#69)\n\n- Move registration and address manager tests to dedicated modules\n- Ensure all required cases are covered and passing\n- Remove duplicate tests from test.rs\n- All tests green (cargo test)
