+++ gateway-contract/contracts/core_contract/src/test.rs
@@ -2,13 +2,18 @@
 
 use crate::{Contract, ContractClient};
 use soroban_sdk::testutils::Address as _;
-use soroban_sdk::{Address, BytesN, Env, String};
+use soroban_sdk::{Address, BytesN, Env};
+
+#[path = "test_registration.rs"]
+mod test_registration;
+#[path = "test_address_manager.rs"]
+mod test_address_manager;
 
 // ============================================
 // Test Setup Helpers
 // ============================================
 
-fn setup_test(env: &Env) -> (ContractClient<'_>, BytesN<32>, Address) {
+pub(crate) fn setup_test(env: &Env) -> (ContractClient<'_>, BytesN<32>, Address) {
     let contract_id = env.register(Contract, ());
     let client = ContractClient::new(env, &contract_id);
     let commitment = BytesN::from_array(env, &[7u8; 32]);
@@ -17,7 +22,7 @@ fn setup_test(env: &Env) -> (ContractClient<'_>, BytesN<32>, Address) {
     (client, commitment, wallet)
 }
 
-fn setup_with_owner(env: &Env) -> (ContractClient<'_>, Address, BytesN<32>) {
+pub(crate) fn setup_with_owner(env: &Env) -> (ContractClient<'_>, Address, BytesN<32>) {
     env.mock_all_auths();
 
     let contract_id = env.register(Contract, ());
@@ -60,249 +65,3 @@ fn test_set_memo_and_resolve_flow() {
     assert_eq!(memo, Some(4242u64));
 }
 
-// ============================================
-// Registration Tests
-// ============================================
-
-#[test]
-fn test_register_success() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let caller = Address::generate(&env);
-    let commitment = BytesN::from_array(&env, &[1u8; 32]);
-
-    // Register should succeed
-    client.register(&caller, &commitment);
-
-    // Verify owner is set
-    let owner = client.get_owner(&commitment);
-    assert_eq!(owner, Some(caller));
-}
-
-#[test]
-#[should_panic(expected = "Commitment already registered")]
-fn test_register_duplicate_rejection() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let caller = Address::generate(&env);
-    let commitment = BytesN::from_array(&env, &[2u8; 32]);
-
-    // First registration succeeds
-    client.register(&caller, &commitment);
-
-    // Second registration with same commitment should panic
-    client.register(&caller, &commitment);
-}
-
-#[test]
-fn test_get_owner_returns_owner_after_registration() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let caller = Address::generate(&env);
-    let commitment = BytesN::from_array(&env, &[3u8; 32]);
-
-    client.register(&caller, &commitment);
-
-    let owner = client.get_owner(&commitment);
-    assert_eq!(owner, Some(caller));
-}
-
-#[test]
-fn test_get_owner_returns_none_for_unknown() {
-    let env = Env::default();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let unknown_commitment = BytesN::from_array(&env, &[99u8; 32]);
-
-    let owner = client.get_owner(&unknown_commitment);
-    assert_eq!(owner, None);
-}
-
-// ============================================
-// Address Manager Tests
-// ============================================
-
-#[test]
-fn test_set_master_stellar_address_success() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let stellar_address = String::from_str(&env, "GXXXXXXXXXXXXXXX");
-
-    // Register address first (required for set_master)
-    client.register_address(&owner, &owner);
-
-    // Set master address
-    client.set_master_stellar_address(&owner, &commitment, &stellar_address);
-
-    // Verify master is set
-    let master = client.get_master();
-    assert_eq!(master, Some(stellar_address));
-}
-
-#[test]
-#[should_panic(expected = "Address not registered")]
-fn test_set_master_stellar_address_not_registered() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let stellar_address = String::from_str(&env, "GXXXXXXXXXXXXXXX");
-
-    // Don't register address - should fail
-    client.set_master_stellar_address(&owner, &commitment, &stellar_address);
-}
-
-#[test]
-#[should_panic(expected = "Not owner")]
-fn test_set_master_stellar_address_non_owner() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let non_owner = Address::generate(&env);
-    let stellar_address = String::from_str(&env, "GXXXXXXXXXXXXXXX");
-
-    // Register address
-    client.register_address(&owner, &owner);
-
-    // Non-owner tries to set master - should fail
-    client.set_master_stellar_address(&non_owner, &commitment, &stellar_address);
-}
-
-#[test]
-fn test_add_stellar_address_success() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let stellar_address = String::from_str(&env, "GYYYYYYYYYYYYYYYY");
-
-    // Add stellar address
-    client.add_stellar_address(&owner, &commitment, &stellar_address);
-
-    // Verify address is stored
-    let stored = client.get_stellar_address(&commitment);
-    assert_eq!(stored, Some(stellar_address));
-}
-
-#[test]
-#[should_panic(expected = "Address already exists")]
-fn test_add_stellar_address_duplicate_rejection() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let stellar_address = String::from_str(&env, "GZZZZZZZZZZZZZZZZ");
-
-    // First add succeeds
-    client.add_stellar_address(&owner, &commitment, &stellar_address);
-
-    // Second add with same commitment should panic
-    client.add_stellar_address(&owner, &commitment, &stellar_address);
-}
-
-#[test]
-#[should_panic(expected = "Not owner")]
-fn test_add_stellar_address_non_owner_rejection() {
-    let env = Env::default();
-    let (client, _owner, commitment) = setup_with_owner(&env);
-    let non_owner = Address::generate(&env);
-    let stellar_address = String::from_str(&env, "GAAAAAAAAAAAAAAAA");
-
-    // Non-owner tries to add - should fail
-    client.add_stellar_address(&non_owner, &commitment, &stellar_address);
-}
-
-#[test]
-fn test_register_address_success() {
-    let env = Env::default();
-    let (client, owner, _commitment) = setup_with_owner(&env);
-    let address_to_register = Address::generate(&env);
-
-    // Register address
-    client.register_address(&owner, &address_to_register);
-
-    // Verify address is registered
-    let is_registered = client.is_address_registered(&address_to_register);
-    assert!(is_registered);
-}
-
-#[test]
-#[should_panic(expected = "Not owner")]
-fn test_register_address_non_owner_rejection() {
-    let env = Env::default();
-    let (client, _owner, _commitment) = setup_with_owner(&env);
-    let non_owner = Address::generate(&env);
-    let address_to_register = Address::generate(&env);
-
-    // Non-owner tries to register - should fail
-    client.register_address(&non_owner, &address_to_register);
-}
-
-#[test]
-#[should_panic(expected = "Address already registered")]
-fn test_register_address_duplicate_rejection() {
-    let env = Env::default();
-    let (client, owner, _commitment) = setup_with_owner(&env);
-    let address_to_register = Address::generate(&env);
-
-    // First registration succeeds
-    client.register_address(&owner, &address_to_register);
-
-    // Second registration should fail
-    client.register_address(&owner, &address_to_register);
-}
-
-#[test]
-fn test_get_master_returns_correct_address() {
-    let env = Env::default();
-    let (client, owner, commitment) = setup_with_owner(&env);
-    let stellar_address = String::from_str(&env, "GBBBBBBBBBBBBBBB");
-
-    // Register and set master
-    client.register_address(&owner, &owner);
-    client.set_master_stellar_address(&owner, &commitment, &stellar_address);
-
-    // Verify get_master returns correct address
-    let master = client.get_master();
-    assert_eq!(master, Some(stellar_address));
-}
-
-#[test]
-fn test_get_master_returns_none_when_not_set() {
-    let env = Env::default();
-    let (client, _owner, _commitment) = setup_with_owner(&env);
-
-    // Don't set master
-    let master = client.get_master();
-    assert_eq!(master, None);
-}
-
-#[test]
-fn test_is_address_registered_false_for_unknown() {
-    let env = Env::default();
-    let (client, _owner, _commitment) = setup_with_owner(&env);
-    let unknown_address = Address::generate(&env);
-
-    let is_registered = client.is_address_registered(&unknown_address);
-    assert!(!is_registered);
-}
-
-#[test]
-#[should_panic(expected = "Already initialized")]
-fn test_init_address_manager_twice_fails() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(Contract, ());
-    let client = ContractClient::new(&env, &contract_id);
-    let owner = Address::generate(&env);
-
-    // First init succeeds
-    client.init_address_manager(&owner);
-
-    // Second init should fail
-    client.init_address_manager(&owner);
-}
